% Auto-generated: do not edit by hand
\name{dialogBody}

\alias{dialogBody}

\title{DialogBody component}

\description{
Body of the dialog, optionally with a constrained container height which allows vertical scrolling of its content.
}

\usage{
dialogBody(children=NULL, id=NULL, className=NULL,
useOverflowScrollContainer=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. DialogBody contents.}

\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{className}{Character. A space-delimited list of class names to pass along to a child element.}

\item{useOverflowScrollContainer}{Logical. Enable scrolling for the container}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

