% Auto-generated: do not edit by hand
\name{divider}

\alias{divider}

\title{Divider component}

\description{
Divider visually separate contents with a thin line and margin on all sides. Dividers work best  in flex layouts where they will adapt to orientation without additional styles. Otherwise, a  divider will appear as a full-width 1px-high block element.
}

\usage{
divider(className=NULL, tagName=NULL)
}

\arguments{
\item{className}{Character. A space-delimited list of class names to pass along to a child element.}

\item{tagName}{HTML tag to use for element.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

