% Auto-generated: do not edit by hand
\name{menu}

\alias{menu}

\title{Menu component}

\description{
Menus display lists of interactive items.
}

\usage{
menu(children=NULL, id=NULL, className=NULL, large=NULL,
small=NULL, style=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. Menu items.}

\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{className}{Character. A space-delimited list of class names to pass along to a child element.}

\item{large}{Logical. Whether the menu items in this menu should use a large appearance.}

\item{small}{Logical. Whether the menu items in this menu should use a small appearance.}

\item{style}{Named list. CSS properties to apply to the menu.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

