% Auto-generated: do not edit by hand
\name{navbarDivider}

\alias{navbarDivider}

\title{NavbarDivider component}

\description{
Use to divide components of the navbar
}

\usage{
navbarDivider(className=NULL)
}

\arguments{
\item{className}{Character. A space-delimited list of class names to pass along to a child element.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

