% Auto-generated: do not edit by hand
\name{navbarGroup}

\alias{navbarGroup}

\title{NavbarGroup component}

\description{
Use to group components of the navbar
}

\usage{
navbarGroup(children=NULL, id=NULL, align=NULL, className=NULL,
style=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. Content of the NavbarGroup}

\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{align}{Character. The side of the navbar on which the group should appear. The 
Alignment enum provides constants for these values.}

\item{className}{Character. A space-delimited list of class names to pass along to a child element.}

\item{style}{Named list. CSS properties to apply to the root element.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

