% Auto-generated: do not edit by hand
\name{navbarHeading}

\alias{navbarHeading}

\title{NavbarHeading component}

\description{
Heading of the navbar
}

\usage{
navbarHeading(children=NULL, id=NULL, className=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. Content of the NavbarHeading}

\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{className}{Character. A space-delimited list of class names to pass along to a child element.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

