% Auto-generated: do not edit by hand
\name{sectionCard}

\alias{sectionCard}

\title{SectionCard component}

\description{
Multiple SectionCard child components can be added under one Section, they will be stacked vertically.  This layout can be used to further group information.
}

\usage{
sectionCard(children=NULL, id=NULL, className=NULL, padded=NULL,
style=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. Content of the Card}

\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{className}{Character. A space-delimited list of class names to pass along to a child element.}

\item{padded}{Logical. Whether to apply visual padding inside the content container element.}

\item{style}{Named list. CSS properties to apply to the root element.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

