% Auto-generated: do not edit by hand
\name{sideBar}

\alias{sideBar}

\title{SideBar component}

\description{
Component for creating interactive Sidebars
}

\usage{
sideBar(id=NULL, className=NULL, initialRoute=NULL, items=NULL,
route=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{className}{Character. A space-delimited list of class names to pass along to a child element.}

\item{initialRoute}{Character. initial route}

\item{items}{Unnamed list. items to be displayed in the menu.}

\item{route}{Character. current selected route}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

