% Auto-generated: do not edit by hand
\name{menuDivider}

\alias{menuDivider}

\title{MenuDivider component}

\description{
Use MenuDivider to separate menu sections. Optionally, add a title to the divider.
}

\usage{
menuDivider(id=NULL, className=NULL, title=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{className}{Character. A space-delimited list of class names to pass along to a child element.}

\item{title}{Character. Optional header title.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

