% Auto-generated: do not edit by hand
\name{navbar}

\alias{navbar}

\title{Navbar component}

\description{
Navbars present useful navigation controls at the top of an application.
}

\usage{
navbar(children=NULL, id=NULL, className=NULL, fixedToTop=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. Content of the Navbar}

\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{className}{Character. A space-delimited list of class names to pass along to a child element.}

\item{fixedToTop}{Logical. Whether this navbar should be fixed to the top of the viewport 
(using CSS position: fixed)}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

